Move global hooks to `shared` directory
